PYTHON/requirements.txt
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pytesseract==0.3.10
PyMuPDF==1.23.14
pdf2image==3.1.0
opencv-python==4.8.1.78
fpdf2==2.7.6
Pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
langchain-community==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4